src/agl: First take at passing down multiple surfaces
For the moment this brings in the necessary bits to pass down multiple
surfaces. The sender/receiver protocol only allows to pass strings,
so we need to serialize the list of surfaces somehow.

This patch introduces protobuf protocol to be able to send the surfaces
as a list. Passing multiple surfaces as a list is a requirement to be
able to manage simultaneously multiple surfaces, not just one and it is
needed to the client shell in AGL project. Regular application wouldn't
need it.

Bug-AGL: SPEC-3799, SPEC-3961

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>