Create a dynamic timer and use dynamic args for starting up
Not doing so will make other, in-flight requests, to over-write the
timer together with the start-up args, which will in turn has the effect
of missing to start-up for an already application start-up.

This adds an embedded class StartupArgs into WebAppManagerServiceAGL
class, to be able to pass arguments dynamically and avoid the issue
described above. Further more this patch expands the Base Timer class
to allow calling a callback with an argument.

Bug-AGL: SPEC-3969

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>