Introduce kSendReady event
This introduces ready() request handling. Once all surfaces have been
loaded we can signal the compositor to let it know it can start
presenting. We add a new kSendReady event to trigger the request in on
the chromium side.

With this we clean up a bit the surface encoding part, to avoid
it entirely if it is not necessary (when handling any other types of events:
start/stopping versus ready).

Bug-AGL: SPEC-4009

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>