[op][agl] New QtLess AGL port
Single commit including the baseline of AGL port of WAM as it was in
branch @52.agl.marlin, but adapted to upstream QtLess work that
happened between submissions/52 and submissions/55.

This is done in a single commit, as rewriting the history on top of
new QtLess upstream, but skipping our own (now legacy) QtLess would
be a lot of work, and with no clear value. For research of reasons
for a specific commit code, that points to this commit, we recommend
checking in branch @52.agl.marlin (the last one before the upstream
new QtLess).

The new coding standard is following Google C++ Style guide at
https://google.github.io/styleguide/cppguide.html and can be checked
with git-clang-format.

To track the authors of the original commits, see previous WAM history
from @52.agl.marlin branch.

Bug-AGL: 4239
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>