[op][agl][wayland] Properly pass surface size, type and edge to chromium
It was found applications being launched were not properly keeping
the application description data for the type of application, name,
etc, effectively breaking launching background and home screen
applications.

This would cause AGL ready event to not be called for background
application when it was ready, preventing from showing any UI on
screen.

Bug-AGL: SPEC-4250
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>