[op][agl][wayland] Set AGL application ID in all cases
After change to support fully agl_shell logic, we need to know which
application implements the agl_shell interface UI (in demo case
html5-homescreen). So we need to pass the application ID in all cases
and not only when it is a regular application.

In any case, application ID is important for all the task swithing
cases.

Bug-AGL: SPEC_4250
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>